refactor and remove popup
