--- src/pages/popup/index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-  <head>
-    <meta charset="UTF-8" />
-    <title>Popup</title>
-  </head>
-
-  <body>
-    <div id="app-container"></div>
-    <script type="module" src="./index.tsx"></script>
-  </body>
-</html>
