--- src/pages/popup/index.tsx
@@ -1,24 +1,0 @@
-import React from 'react';
-import { createRoot } from 'react-dom/client';
-import '@pages/popup/index.css';
-import Popup from '@pages/popup/Popup';
-import { BrowserRouter as Router } from 'react-router-dom'; // Import BrowserRouter
-// import refreshOnUpdate from 'virtual:reload-on-update-in-view';
-
-// refreshOnUpdate('pages/popup');
-
-function init() {
-  const appContainer = document.querySelector('#app-container');
-  if (!appContainer) {
-    throw new Error('Can not find #app-container');
-  }
-  const root = createRoot(appContainer);
-
-  root.render(
-    <Router>
-      <Popup />
-    </Router>,
-  );
-}
-
-init();
