+++ src/pages/sidepanel/assets/data/sampleConfig.json
