+++ src/pages/sidepanel/assets/svg/developerCenterIcon.tsx
