+++ src/pages/sidepanel/assets/svg/empty.tsx
