+++ src/pages/sidepanel/assets/svg/logo.tsx
