+++ src/pages/sidepanel/assets/svg/wrongDomainIcon.tsx
