+++ src/pages/sidepanel/components/SimpleTable.tsx
