+++ src/pages/sidepanel/sections/Configuration.tsx
