+++ src/shared/storages/base.ts
@@ -62,7 +62,7 @@ type StorageConfig = {
    */
   sessionAccessForContentScripts?: boolean;
   /**
-   * Keeps state live in sync between all instances of the extension. Like between popup, side panel and content scripts.
+   * Keeps state live in sync between all instances of the extension. Like between side panel and content scripts.
    * To allow chrome background scripts to stay in sync as well, use {@link StorageType.Session} storage area with
    * {@link StorageConfig.sessionAccessForContentScripts} potentially also set to true.
    * @see https://stackoverflow.com/a/75637138/2763239
