+++ src/shared/storages/extensionDomainStorage.ts
@@ -22,28 +22,3 @@ const domainStore: DomainStorage = {
 };
 
 export default domainStore;
-
-// import { BaseStorage, createStorage, StorageType } from '@src/shared/storages/base';
-// import { EmitLog } from '@src/shared/components/EmitLog';
-
-// const ExtensionConfigStorageKey = 'lyticsDomain';
-
-// type ExtensionConfigStorage = BaseStorage<string> & {
-//   clear: () => void;
-// };
-
-// const storage = createStorage<string>(ExtensionConfigStorageKey, '', {
-//   storageType: StorageType.Local,
-//   liveUpdate: true,
-// });
-
-// const extensionConfigStorage: ExtensionConfigStorage = {
-//   ...storage,
-//   clear: () => {
-//     chrome.storage.local.remove(ExtensionConfigStorageKey, () => {
-//       EmitLog({ name: 'storage', payload: { msg: `Data associated with ${ExtensionConfigStorageKey} cleared.` } });
-//     });
-//   },
-// };
-
-// export default extensionConfigStorage;
